src/requirements.txt
python-dotenv>=0.19.0
nextcord>=2.0.0
os